main/postmarketos-mkinitfs-hook-liveboot: new aports
liveboot is a ramdisk hook to allow for booting a rootfs entirely from
RAM, allowing users to try out postmarketOS without installing - much in
the same vain as a classic Linux livecd.

Unfortunately, due to the extra complexity of implementing this on a
phone it requires ~3GB of RAM for an average Phosh rootfs, making it
only feasible on a limited number of devices.

It works by creating a 600mb vfat image and offering it to the host
machine via mass storage gadget, the user is then instructed to copy a
standard .img.xz rootfs which they have downloaded from
images.postmarketos.org. The script waits for the user to press volume
down to indicate that the copy has completed. It will then mount the
vfat image, copy out the compressed rootfs and delete the vfat image.

The rootfs is then uncompressed (hence the requirement that there is
enough RAM to fit the compressed and uncompressed rootfs at once) and
mounted before boot continues as normal.